+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 
 The format os based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this package adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased 
+
+### Changed
+
+- Fix double input events
+- Optimize components and mixins
+- Fix focus on fields after reset
+
 ## v0.1.10
 
 ### Deprecated